removed user input for get_t_list()
replaced the requirement for a user input if no tracker.toml file exists or if fails to open and read and load for some reason it will return false.

if tracker.toml contains valid ips it will attempt to connect to them and get their tracker list and add it to its own and outputs and returns it.